# src/main/resources/swagger/api.yaml
openapi: 3.0.1
info:
  title: Ecommerce APIs
  description: Open API specifications for Ecommerce Application
  version: 1.0.0
paths:
...
  /products/{id}:
    get:
      operationId: getProduct
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: productId

      responses:
        200:
          description: success, product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/product'
        404:
          description: product not found
        500:
          description: Internal error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /order/{id}:
    get:
      operationId: getOrder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: orderId
        - in: header
          name: X-API-KEY
          schema:
            type: string
            format: uuid
          required: true

      responses:
        200:
          description: success, order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
        404:
          description: order not found
        500:
          description: Internal error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /order/place:
    post:
      operationId: placeFreshOrder
      parameters:
      - in: query
        name: userId
        required: false
        schema:
          type: integer
      - in: header
        name: X-API-KEY
        schema:
          type: string
          format: uuid
        required: true

      requestBody:
        description: The required body of order
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/order'

      responses:
        200:
          description: success, order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
        500:
          description: Internal error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /order/addProduct/{id}:
    post:
      operationId: addToExistingOrder
      parameters:
        - in: header
          name: X-API-KEY
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: orderId

      requestBody:
        description: The required body of orderProduct
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orderProduct'

      responses:
        200:
          description: success, order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
        500:
          description: Internal error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  schemas:
    order:
      type: object
      required:
        - id
        - name
        - userId
        - totalPrice
        - orderProducts
      properties:
        id:
          type: integer
          nullable: false
        name:
          type: string
          nullable: false
        userId:
          type: integer
          nullable: false
        totalPrice:
          type: number
        orderProducts:
          $ref: '#/components/schemas/orderProducts'
    orderProducts:
      type: array
      items:
        $ref: '#/components/schemas/orderProduct'
    orderProduct:
      type: object
      required:
        - product
        - quantity
      properties:
        product:
          $ref: '#/components/schemas/product'
        quantity:
          type: integer
...
    product:
      type: object
      required:
        - id
        - name
        - description
        - price
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
...
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string